+++ src/types.ts
@@ -1443,6 +1443,7 @@ export type MutationAddTagArgs = {
 
 export type MutationAddTemplateArgs = {
   copyFromTemplateId?: InputMaybe<Scalars['Int']['input']>;
+  copyFromVersionedTemplateId?: InputMaybe<Scalars['Int']['input']>;
   name: Scalars['String']['input'];
 };
 
@@ -3804,6 +3805,8 @@ export type Template = {
   sections?: Maybe<Array<Maybe<Section>>>;
   /** The template that this one was derived from */
   sourceTemplateId?: Maybe<Scalars['Int']['output']>;
+  /** The versioned template that this one was derived from */
+  sourceVersionedTemplateId?: Maybe<Scalars['Int']['output']>;
 };
 
 /** A user that that belongs to a different affiliation that can edit the Template */
@@ -3903,6 +3906,7 @@ export type TemplateErrors = {
   ownerId?: Maybe<Scalars['String']['output']>;
   sectionIds?: Maybe<Scalars['String']['output']>;
   sourceTemplateId?: Maybe<Scalars['String']['output']>;
+  sourceVersionedTemplateId?: Maybe<Scalars['String']['output']>;
 };
 
 /** A search result for templates */
@@ -3928,6 +3932,8 @@ export type TemplateSearchResult = {
   latestPublishVersion?: Maybe<Scalars['String']['output']>;
   /** Visibility set for the last published template */
   latestPublishVisibility?: Maybe<TemplateVisibility>;
+  /** versionedTemplateId of the template */
+  latestVersionedTemplateId?: Maybe<Scalars['Int']['output']>;
   /** The timestamp when the Template was last modified */
   modified?: Maybe<Scalars['String']['output']>;
   /** The id of the person who last modified the template */
@@ -6624,6 +6630,7 @@ export type TemplateResolvers<ContextType = MyContext, ParentType extends Resolv
   owner?: Resolver<Maybe<ResolversTypes['Affiliation']>, ParentType, ContextType>;
   sections?: Resolver<Maybe<Array<Maybe<ResolversTypes['Section']>>>, ParentType, ContextType>;
   sourceTemplateId?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
+  sourceVersionedTemplateId?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
 };
 
 export type TemplateCollaboratorResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['TemplateCollaborator'] = ResolversParentTypes['TemplateCollaborator']> = {
@@ -6673,6 +6680,7 @@ export type TemplateErrorsResolvers<ContextType = MyContext, ParentType extends 
   ownerId?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   sectionIds?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   sourceTemplateId?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  sourceVersionedTemplateId?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
 };
 
 export type TemplateSearchResultResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['TemplateSearchResult'] = ResolversParentTypes['TemplateSearchResult']> = {
@@ -6686,6 +6694,7 @@ export type TemplateSearchResultResolvers<ContextType = MyContext, ParentType ex
   latestPublishDate?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   latestPublishVersion?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   latestPublishVisibility?: Resolver<Maybe<ResolversTypes['TemplateVisibility']>, ParentType, ContextType>;
+  latestVersionedTemplateId?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
   modified?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   modifiedById?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
   modifiedByName?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;